// src/integrations/supabase/types.ts
export type Database = {
  public: {
    Tables: {
// ...
      rooms: {
        Row: {
          capacity: number
          created_at: string | null
          floor: number
          id: string
          price: number
          room_number: string
          room_type: string
          status: string
          tenant_id: string | null
          updated_at: string | null
        }
        Insert: {
          capacity: number
          created_at?: string | null
          floor: number
          id?: string
          price: number
          room_number: string
          room_type: string
          status?: string
          tenant_id?: string | null
          updated_at?: string | null
        }
        Update: {
          capacity?: number
          created_at?: string | null
          floor?: number
          id?: string
          price?: number
          room_number?: string
          room_type?: string
          status?: string
          tenant_id?: string | null
          updated_at?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "rooms_tenants_fkey"
            columns: ["tenants"]
            isOneToOne: false
            referencedRelation: "tenants"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      is_admin: {
        Args: Record<PropertyKey, never>
        Returns: boolean
      }
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}